Add new features to Collection
Add the ability to execute a command on each model in a collection -
`each()`.

Add the ability to filter collections through the `find` method. Add
some shorthand for common taks using the filter command:

 * Getting the first object which matches
 * Checking if at least one object matches

`find` returns a Collection, so chaining of commands is possible.

All methods take a callback function which should accept one param,
the model to execute on. `find`, `find_one`, and `contains` should
return TRUE or FALSE to specify if the model is considered to match
the query.